backend/requirements.txt
langchain-core
langchain-groq
crewai
crewai-tools
pydantic
flask
flask-cors
supabase
python-jose
astrapy
langchain 
langchain-astradb
python-dotenv
python-docx
pymupdf